added more plots and main now produces nice dashboard of recent results

+++ lib/cleaning/z_norm.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 namespace z_norm {
-  void z_normalise(std::vector<float>& series);
+  void z_normalise(std::vector<double>& series);
 };
 
 #endif
