+++ lib/dimension_reductions/paa.cpp
@@ -2,10 +2,10 @@
 
 using std::vector;
 
-float paa::get_mean(const float* const first, const float* const last)
+double paa::get_mean(const double* const first, const double* const last)
 {
   if (first > last) return 0.0;
-  float sum = 0; 
+  double sum = 0; 
 
   for (int i=0; i<=last-first; ++i) {
     sum += *(first+i);
@@ -13,11 +13,11 @@ float paa::get_mean(const float* const first, const float* const last)
   return sum / (last - first + 1);
 }
 
-vector<float> paa::paa(const vector<float>& series, unsigned int interval_size)
+vector<double> paa::paa(const vector<double>& series, unsigned int interval_size)
 {
   if (series.size() == 0) return {};
 
-  vector<float> paa_vec;
+  vector<double> paa_vec;
 
   int num_full_intervals = (series.size() / interval_size);
   for (int i=0; i<num_full_intervals; ++i) {
@@ -34,12 +34,12 @@ vector<float> paa::paa(const vector<float>& series, unsigned int interval_size)
   return paa_vec;
 }
 
-float paa::paa_mse(const vector<float>& series, unsigned int interval_size)
+double paa::paa_mse(const vector<double>& series, unsigned int interval_size)
 {
-  vector<float> paa_series = paa::paa(series, interval_size);
-  float mse = 0;
+  vector<double> paa_series = paa::paa(series, interval_size);
+  double mse = 0;
   for (int i=0; i<series.size(); ++i) {
-    float diff = (series[i] - paa_series[ i / interval_size ]);
+    double diff = (series[i] - paa_series[ i / interval_size ]);
     mse += diff * diff;
   }
   return mse / series.size();
